message_feed: Show names of reacting users if total votes are 3 or less.
This commit implements the user setting to show the names of users
who reacted on a message. When this setting is turned on, it shows
the names of the reacting users if total votes on the message are
less than or equal to 3; and reverts back to showing the count if
the total votes exceed 3.

Fixes: #20980.